transit: Sort NearbyTransports by distance, logs++
Changed around how 'services' and 'sources' work; sources are
internal resources used to provide data to 'services', which will
manipulate the raw data and return the final result to the user.

Sources may or may not have logging; services should always have
logging.

If a service is found performing the exact same methods as a source,
consider removing the concept of a source for that particular component,
and have the 'source package' directly return a service (and implement
logging in the 'source package').